Moved metrics to sc-external
