+++ external/src/main/java/com/digitalpebble/storm/crawler/metrics/TimerMetric.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.digitalpebble.storm.metrics;
+package com.digitalpebble.storm.crawler.metrics;
 
 import java.util.HashMap;
 import java.util.Map;
